pyproject.toml
[tool.poetry]
name = "paste-server-python"
version = "0.1.0"
description = ""
authors = ["tpob <tpob@tpob.xyz>"]

[tool.poetry.dependencies]
python = "^3.9"
httpx = "^0.20.0"
fastapi = "^0.70.0"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
pycryptodome = "^3.11.0"
websockets = "^10.0"

[tool.poetry.dev-dependencies]
black = "^21.10b0"
ipython = "^7.29.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
